Fix Scala Steward by allowing building under Java 21
Scala Steward can't run on this repo until it allows building under Java 21, due to
https://github.com/guardian/scala-steward-public-repos/pull/67 ... we're now running
our Scala Steward workflow with Java 21 LTS, which means that all projects that want
to have Scala Steward updates (like this one) need to be able to build under Java 21
(even if the projects are still _running_ in production on Java 8).

Being able to build with Java 21 [requires](https://docs.scala-lang.org/overviews/jdk-compatibility/overview.html#jdk-21-compatibility-notes)
using recent versions of Scala (2.12.18+) and sbt (1.9.0+) - for this repo, all that's
needed is a bump of sbt version!